+++ src/lib/remarkIndexedBlock.ts
@@ -1,4 +1,4 @@
-import { Root, Paragraph, Content } from "mdast";
+import { Root, Content } from "mdast";
 import { Plugin } from "unified";
 
 export type IndexedBlock = {
@@ -13,11 +13,18 @@ declare module "mdast" {
   }
 }
 
-type Options = {};
+type Options = {
+  fileName: string;
+  idGenerator?: (index: number, fileName: string) => string;
+};
 
 const key = "st1the";
 
-export const remarkIndexedBlock: Plugin<[], Root> = () => {
+export const remarkIndexedBlock: Plugin<[Options], Root> = (options) => {
+  if (!options.fileName) {
+    throw Error("filename option is requried");
+  }
+
   return (tree) => {
     const newTree: Content[] = [];
     for (const child of tree.children) {
@@ -25,7 +32,9 @@ export const remarkIndexedBlock: Plugin<[], Root> = () => {
         const index = tree.children.indexOf(child);
         const indexedBlockId: IndexedBlock = {
           type: "indexedBlock",
-          id: `${key}-${index}`,
+          id: options.idGenerator
+            ? options.idGenerator(index, options.fileName)
+            : `${key}-${index}`,
           children: [child],
         };
         newTree.push(indexedBlockId);
